test(cmd/bd): build-tag cgo-only test helpers so pure-go tests compile
The cmd/bd test package was wedged: a handful of test files (init_test.go,
bootstrap_test.go, test_helpers_test.go) carry //go:build cgo because they
pull in internal/storage/dolt, database/sql, and the embedded Dolt server.
Their helpers — captureStdout, captureStderr, stdioMutex, ensureTestMode,
saveAndRestoreGlobals, buildBDForInitTests, runGitForBootstrapTest, and
friends — are pure Go in body but were stranded behind the cgo tag along
with the heavy ones.

Untagged test files (dolt_test.go, init_safety_test.go, where_test.go)
called those helpers directly, so the entire cmd/bd test binary failed to
compile under CGO_ENABLED=0 with the gms_pure_go tag:

    cmd/bd/dolt_test.go:1237:10: undefined: captureStdout
    cmd/bd/init_safety_test.go:230:11: undefined: buildBDForInitTests
    cmd/bd/init_safety_test.go:233:2: undefined: runGitForBootstrapTest
    ...

PR #3704 (CLI docs from live help) added cmd/bd/help_all_test.go with
TestHelpList/TestHelpAll/TestHelpDoc — a pure-Go test that should run
without cgo. With cmd/bd's test binary uncompilable under CGO_ENABLED=0,
those tests were silent: codecov dropped patch coverage to 75% partly
because the new tests literally could not run via go test.

Fix: extract the pure-Go subset of test helpers into a new untagged file
test_helpers_pure_test.go and remove the duplicate definitions from the
cgo-tagged files. The cmd/bd test binary now compiles cleanly under
CGO_ENABLED=0 -tags gms_pure_go, and untagged test files no longer need
their own //go:build cgo because the helpers they call are available
without cgo.

CI lock-in: a new check-cmd-bd-puregeo-tests job in .github/workflows/ci.yml
compiles cmd/bd with CGO_ENABLED=0 and runs the pure-Go test subset, so a
regression that re-introduces cgo coupling fails fast at PR time rather
than silently neutering pure-Go tests.

Refs: mybd-ycx, GH#3683, PR #3704
