+++ cmd/bd/init_test.go
@@ -11,7 +11,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/steveyegge/beads/internal/beads"
@@ -607,31 +606,6 @@ func TestSetupGlobalGitIgnore_ReadOnly(t *testing.T) {
 			t.Error("expected .beads pattern in output")
 		}
 	})
-}
-
-// captureStdout captures stdout output from fn and returns it as a string.
-// Uses stdioMutex to prevent races with concurrent os.Stdout redirection (bd-cqjoi).
-func captureStdout(t *testing.T, fn func() error) string {
-	t.Helper()
-
-	stdioMutex.Lock()
-	defer stdioMutex.Unlock()
-
-	oldStdout := os.Stdout
-	r, w, _ := os.Pipe()
-	os.Stdout = w
-
-	err := fn()
-
-	w.Close()
-	var buf bytes.Buffer
-	buf.ReadFrom(r)
-	os.Stdout = oldStdout
-
-	if err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-	return buf.String()
 }
 
 // TestInitPromptRoleConfig tests the beads.role git config read/write functions
@@ -1667,46 +1641,6 @@ func TestInitDoltMetadataNoGit(t *testing.T) {
 	if _, err := os.Stat(sqlitePath); err == nil {
 		t.Errorf("unexpected sqlite database created in dolt mode")
 	}
-}
-
-// buildBDOnce builds the bd binary once for subprocess tests in this file.
-// Uses sync.Once for efficiency when multiple tests need the binary.
-var (
-	initTestBD     string
-	initTestBDOnce sync.Once
-	initTestBDErr  error
-)
-
-func buildBDForInitTests(t *testing.T) string {
-	t.Helper()
-	initTestBDOnce.Do(func() {
-		// Check if bd binary exists in repo root (../../bd from cmd/bd/)
-		bdBinary := "bd"
-		if runtime.GOOS == "windows" {
-			bdBinary = "bd.exe"
-		}
-		repoRoot := filepath.Join("..", "..")
-		existingBD := filepath.Join(repoRoot, bdBinary)
-		if _, err := os.Stat(existingBD); err == nil {
-			initTestBD, _ = filepath.Abs(existingBD)
-			return
-		}
-		// Fall back to building
-		tmpDir, err := os.MkdirTemp("", "bd-init-test-*")
-		if err != nil {
-			initTestBDErr = fmt.Errorf("failed to create temp dir: %w", err)
-			return
-		}
-		initTestBD = filepath.Join(tmpDir, bdBinary)
-		cmd := exec.Command("go", "build", "-tags", "gms_pure_go", "-o", initTestBD, ".")
-		if out, err := cmd.CombinedOutput(); err != nil {
-			initTestBDErr = fmt.Errorf("go build failed: %v\n%s", err, out)
-		}
-	})
-	if initTestBDErr != nil {
-		t.Fatalf("Failed to build bd binary: %v", initTestBDErr)
-	}
-	return initTestBD
 }
 
 func setupBareParentInitWorktree(t *testing.T) (string, string) {
